pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nexa-api"
version = "0.3.0"
description = "Nexa API - Remote AI Pipeline Service"
authors = [
  { name="Nexa Team", email="team@nexa.run" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "accelerate>=0.25.0",
    "dask>=2024.1.0",
    "datasets>=2.14.0",
    "distributed>=2024.1.0",
    "fastapi>=0.115",
    "graphene>=3.3.0",
    "h5py>=3.10.0",
    "mlflow>=2.8",
    "numpy>=1.26",
    "nvidia-ml-py>=12.535",
    "openai>=1.0.0",
    "pandas>=2.1",
    "pint>=0.24",
    "prometheus-client>=0.20.0",
    "psutil>=5.9.0",
    "pyarrow>=14.0.0",
    "pybloom-live>=4.0.0",
    "pydantic>=2.4,<2.12",
    "pydantic-settings>=2.0.0,<2.2",
    "pyyaml>=6.0",
    "qdrant-client>=1.7.0",
    "questionary>=2.0.0",
    "requests",
    "rich>=13.6",
    "scikit-learn>=1.3",
    "scipy>=1.11.0",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.28.0",
    "tensorboard>=2.14",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "torchvision>=0.16.0",
    "tqdm>=4.66",
    "transformers>=4.35.0",
    "typer>=0.9",
    "uvicorn>=0.30",
    "wandb>=0.16.0",
    "zstandard>=0.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.24.0",  # For FastAPI TestClient
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML",
    "types-requests",
    "pre-commit>=3.5.0",
]

[project.scripts]
nexa = "nexa_compute.cli.main:app"

[project.urls]
"Homepage" = "https://github.com/nexa/nexa-compute"
"Bug Tracker" = "https://github.com/nexa/nexa-compute/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
